feat(pie): move stories to the dedicated storybook workspace
