+++ storybook/package.json
@@ -28,6 +28,7 @@
     "@nivo/line": "workspace:*",
     "@nivo/marimekko": "workspace:*",
     "@nivo/network": "workspace:*",
+    "@nivo/pie": "workspace:*",
     "@nivo/tooltip": "workspace:*",
     "@nivo/waffle": "workspace:*",
     "@babel/preset-env": "^7.21.5",
